+++ types/imagebitmap.js
@@ -1,6 +1,5 @@
 /* globals document, OffscreenCanvas, createImageBitmap -- Polyfills */
-// `ImageBitmap` is browser / DOM specific. It also can only work
-//  same-domain (or CORS)
+// `ImageBitmap` can only work same-domain (or CORS) in the browser
 
 import {toStringTag, TypesonPromise} from 'typeson';
 
@@ -10,10 +9,7 @@ import {toStringTag, TypesonPromise} from 'typeson';
 const imagebitmap = {
     imagebitmap: {
         test (x) {
-            return toStringTag(x) === 'ImageBitmap' ||
-                // In Node, our polyfill sets the dataset on a canvas
-                //  element as JSDom no longer allows overriding toStringTag
-                (x && x.dataset && x.dataset.toStringTag === 'ImageBitmap');
+            return toStringTag(x) === 'ImageBitmap';
         },
         replace (bm) {
             const canvas = document.createElement('canvas');
@@ -27,8 +23,8 @@ const imagebitmap = {
         },
         revive (o) {
             const canvas = typeof OffscreenCanvas === 'undefined'
+                /* c8 ignore next -- Older environments */
                 ? document.createElement('canvas')
-                /* c8 ignore next -- Browser only */
                 : new OffscreenCanvas(o.width, o.height);
             /*
             var req = new XMLHttpRequest();
@@ -45,14 +41,19 @@ const imagebitmap = {
             const img = document.createElement('img');
             // The onload is needed by some browsers per https://stackoverflow.com/a/4776378/271577
             img.addEventListener('load', function () {
-                ctx.drawImage(img, 0, 0);
+                try {
+                    ctx.drawImage(img, 0, 0);
+                } catch {
+                    // Issues on Node
+                }
             });
             img.src = o.dataURL;
+
             // Works in contexts allowing an `ImageBitmap` (We might use
             //   `OffscreenCanvas.transferToBitmap` when supported)
             return typeof OffscreenCanvas === 'undefined'
+                /* c8 ignore next 3 -- Older environments */
                 ? canvas
-                /* c8 ignore next 3 -- Browser only */
                 : /** @type {OffscreenCanvas} */ (
                     canvas
                 ).transferToImageBitmap();